Support arbitrary collections of cache keys (#378)
This ensures that passing a set of keys will work. While Django
supports arbitrary iterables, I've opted not to support those here
as we'd then need to handle cases where we collapse a generator.
I suspect that that could be handled (by passing the resulting
values to Django), but is a bigger change than needed to fix the
original issue.